requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
dj-stripe==2.8.3
Django==5.0.2
django-appconf==1.0.6
django-bootstrap3==23.6
django-bootstrap4==24.1
django-bootstrap5==23.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==23.5
django-imagekit==5.0.0
django-jsonfield==1.4.1
django-phone-field==1.8.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-storages==1.14.2
django-stripe-payments==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environs==10.3.0
gunicorn==21.2.0
httpie==3.2.2
idna==3.6
install==1.3.5
Markdown==3.5.2
markdown-it-py==3.0.0
marshmallow==3.21.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
packaging==23.2
pilkit==3.0
pillow==10.2.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
